Factor out interval formatting for JSON in a new object
